Functional but long runs may error
This does splitting and farming out of the checks to a worker
pool via a buffered channel, but there is a likelihood that on
long runs the buffered channel will overflow. It is counter to the
goal of the tool to analyze the disk at such a granular level, but
something better than a golang deadlock-exit oculd be done.